Modifications to Correction System for Performance Gains, Development of Data Containers for Object Localization (#11)
* John Edits

* Robot Parsing Correctly

* debug

* Debug

* Debugging on Branch John

* Debug

* Debug Changes on X_Drive

* Release Version 0.1.1

* Construction of Mechanical Components for Usage

* Development on Branch John

* Development on Branch John

* Development on Branch John

* Development on Branch John

* Adding Bull2 SmallBot Configuration

* Debug and Testing Tank Drive

* Debug

* Debug

* Debug

* Debug

* Debug

* Debug

* Debug

* Debug

* Debug

* Debug

* Debug

* Debug

* Debug

* Debug

* Debug

* Debug

* Debug

* Debug

* Building TerriBull BULL2: Left Nut

* Build for Mechanical Systems: Tank Drive, Catapult Shooter, Intakes, Rollers, Controller Inputs, Tasking Development & Testing

* Development of X_Drive, Controller Classes for Drivers, Modificaiton of internal program for improvements and bug fixes.

* DriveTask Modification

* DriveTask Modification

* Modifications to Internal Representations of Data

* Create comms-main.cpp

* Fix Build Errors for GameObjects

* Debug Drivebase

* fix

* fix

* remapping based in inches

* Testing Drive

* good changes

* modify approach angle test

* fixing rollers

* fixing rollers

* mod drive code

* fixing drive code

* debug drive

* debug drive

* debug drive

* debug roller

* debug drive

* debug drive

* debug drive

* debug drive

* debug drive

* Serial Update

* Serial

* Debug/Adding Voltage Curves <untested>

* Quick Build for Ryan's Shooter

* Ryan's Shooter Testing on Button (a)

* Merge Branch to Main

* Merge main for testing on [Branch Moon]

* Updating Shooter Development for branch [TheMoon]

---------

+++ include/Controllers/TaskManager/TaskManager.hpp
@@ -53,7 +53,7 @@ public:
 
     }
 
-    void run(int delta) {
+    void run(float delta) {
         if (this->currentTaskSet != nullptr)
         { /* Update and chek our current task*/
         pros::lcd::set_text(7, "Task Number: " + std::to_string(this->tasksCompleted+1));
